Refactor in wrapper to make style more consistent with core
