+++ programs/wrapper/src/processors/deposit.rs
@@ -15,9 +15,9 @@ use solana_program::{
     pubkey::Pubkey,
 };
 
-use super::shared::{
-    check_signer, get_trader_index_hint_for_market, sync, WrapperStateAccountInfo,
-};
+use crate::loader::{check_signer, WrapperStateAccountInfo};
+
+use super::shared::{get_trader_index_hint_for_market, sync};
 
 #[derive(BorshDeserialize, BorshSerialize)]
 pub struct WrapperDepositParams {
@@ -35,7 +35,6 @@ pub(crate) fn process_deposit(
     accounts: &[AccountInfo],
     data: &[u8],
 ) -> ProgramResult {
-    // Load account infos.
     let account_iter: &mut std::slice::Iter<AccountInfo> = &mut accounts.iter();
     let manifest_program: Program =
         Program::new(next_account_info(account_iter)?, &manifest::id())?;
@@ -67,7 +66,7 @@ pub(crate) fn process_deposit(
     let trader_index_hint: Option<DataIndex> =
         get_trader_index_hint_for_market(&wrapper_state, &market.info.key)?;
 
-    // Call the deposit CPI
+    // Call the deposit CPI.
     invoke(
         &deposit_instruction(
             market.key,
